ADD STYLE TO THE USER-INFO CONTAINER
I have styled thenuser-info container in the header to better align the notification bell and the user icon to be symmetrical to the search input. I have further used grid for the layour of the user-info. I have also used the <hr> tag on the paragraghs on the Announcements container to create the underline after the first two paragraphs.